+++ web/app/store.js
@@ -11,13 +11,15 @@ import menu from './stores/menu/menu.reducer';
 import applicationNotifications from '../../src/stores/application-notifications/application-notifications.reducer';
 import fileUploadForm from './stores/file-upload-form/file-upload-form.reducer';
 import buttonExampleForm from './stores/button-example-form/button-example-form.reducer';
+import dynamicallyLoadingComponents from './stores/dynamically-loading-components/dynamically-loading-components.reducer';
 
 let myReducers = {
   preventDoubleClick,
   menu,
   applicationNotifications,
   fileUploadForm,
-  buttonExampleForm
+  buttonExampleForm,
+  dynamicallyLoadingComponents
 };
 
 let reducers = {};