+++ web/app/stores/dynamically-loading-components/dynamically-loading-components.actions.js
@@ -1,0 +1,9 @@
+import * as constants from './dynamically-loading-components.constants';
+
+export const set = (dataSetName, dataSet) => {
+  return {
+    type: constants.SET,
+    dataSetName,
+    dataSet
+  };
+};